gui-io: invoke JTextArea.append() from the EDT
The method JTextArea.append() was thread-safe until Java 7, so we simply
called it from the main thread. Now that we don't have the guarantee of
safety, we must call it from the Swing Event Dispatcher Thread, by using the
SwingUtilities.invokeLater() method and a custom class that implements
Runnable.

We did not use an anonymous inner class because these classes can only access
final attributes of the class, and the message parameter is clearly not final.

Closes #10.