Add latex preamble to use Unicode math symbols
Given that we need some math symbols, the best way to do it for both the HTML
and the PDF versions is to use Unicode. To avoid depending on newer versions
of Sphinx, or fancy LaTeX packages, I am just adding this small preamble to
the generated LaTeX files to interpret the 3 symbols we need as proper LaTeX
math symbols.