Add a default configuration for the FPU
The FPU stores some configuration data inside a properties file, and the
current ways of accessing the properties file assumes that the simulator is
running from a JAR file. This is not always the case (e.g., while running unit
tests), so this commit adds a default configuration for the FPU, so that FPU
unit tests can be performed.

Filed Issue #25.