Cleaner, typed and persistent store for preferences
This commit introduces a cleaner and typed persistent store for user
preferences based on java.util.prefs.Preferences. It replaces the old one,
that used ugly os-dependent code to find a suitable place to store them.

The store will be persistent across different versions and JAR files, a thing
that the old one didn't guarantee because preferences were stored inside the
JAR itself under Linux.

There is some code duplication in the store, but I am not sure if that can be
avoided.