Fix GUICycles window broken visualization
The GUICycles window was broken. Doing a binary search, I found that it has
been broken since I merged the FPU code, and I don't know if the breakage was
already in the code to merge or if I introduced it.

This fix is just am ugly patch, the code that handles this feature is just
too convoluted. If I had infinite time and willpower, I would rewrite it.
However, both my willpower and my time are very limited, so I will just add
this tiny bit of logic that fixes the problems for all the cases that I have
observed.

Automatic testing will be added to detect wrong state transitions during unit
tests; while this does not prove correctness, it will help me to catch
regressions.