Basic structure of a (non-working) GWT toy client
There are still many things to fix in the code itself, see 'ant gwtc' for a
complete list of things to fix.

This commit adds the appropriate Ant targets, the GWT XML file, scaffolding
HTML code and the Java source file for the frontend (which is just an alert()
for now).

Committing this code mostly for not keeping it in my local machine, it's
unusable as of now (but the normal Ant targets should not be broken).