Rename FileReader to FileUtils
And also add some new methods to be used in IOManager.

The interaction with the filesystem won't only be reading files, but also
opening them in IOManager and doing some other metadata operation. It's worth
having a single class to do this (and to reimplement for GWT). Moreover, moved
it to the utils package.