Fix .gitignore
By mistake, in b5bfa98, an IDEA file that should be tracked was removed from
the repo (or at least added in .gitignore). This caused breakage in the
configuration.

Therefore, let's configure properly .gitignore, by using the templates
provided in the GitHub repository github/gitignore, checking the IntelliJ
documentation and documenting the additional .gitignore directives that are
specific to EduMIPS64.

The templates used for .gitignore are: Java, Gradle and JetBrains, and they
were copied as of commit aad5932.

Most notably, this will cause edumips64.iml to be tracked again.