Handle UI update dispatch via a separate thread
This avoids queuing too many updates for the event dispatch thread via
invokeLater(), but also avoids slowing down processing by directly using
invokeAndWait() in the worker.

The benefits w.r.t. invokeAndWait() seem small and inconsistent, to be
honest, but in one case the improvement was of ~10 cycles/sec (on a
speed of 50 cycles/sec). The code is not too complex so I am going for
this approach.