Implement exception-less RAW conflicts
Profiling revealed that a lot of time in unit tests is spent throwing
and catching the RAWException exception across method boundaries.

Apparently, throwing the exception within the method is much cheaper,
hence the changes in this commit.

In this commit, the Instruction.ID() method returns true when a RAW
conflict is detected, false otherwise. While this is not a clean
interface, it makes the longest unit test (testSetBitSort) roughly
30% faster.

Also, while changing the instructions, some of the simplifications
suggested by IntelliJ were applied.

This should mitigate in part Issue #132.