Don't increment cycle counter in the last cycle
This is a small hack to fix issue #48.

Fixing it in a more radical way requires a deep restructuring of the
code, which I almost finished. The restructuring is invasive, and I
don't think it's beneficial to do it, as it might introduce subtle bugs
I do not find now.

Therefore, I am choosing to not increment the cycles counter in the last
cycle, explaining it in the manual. The explanation makes sense, but the
end result is not very elegant, as when the user executes the very last
cycle they won't see the cycles counter increasing.

Also fix all tests with hardcoded number of cycles, decrementing them by
one cycle.