Throw IntegerOverflowException for large immediate
Since we increased the memory limit, we can come across
situations in which a memory label does not fit the immediate
field of ALU I-type instructions.

In this case, instead of throwing a generic exception, throw
a proper synchronous exception of type IntegerOverflowException.

Also add a unit test for this new behavior.

Fixes #255