Do not incorrectly overwrite EX if FP was executed (#305)
* Improve debuggability of FPPipeline

By adding intermediate variables.

* Add unit test for issue 304

For now address only the functional part, the display part will be
addressed in another test.

* Do not override EX after RAW if FP was executed

This should fix part of #308.

In presence of a RAW stall detected in an ID stage, the EX stage was by
default filled with a BUBBLE (empty instruction). This did actually
overwrite an existing instruction if the instruction that went to MEM happens
to come from the FPU.

* Ignore .s files in project root

* Fix incorrect cycles in unit test for 304

Unfortunately this doesn't yet fix bug 304.

* Add timeouts for tests

* Raise test timeouts from 1s to 2s

This will account for slowness in the CI pipeline.

* Raise large test timeouts to 20s

This allows them to complete successfully on my old Macbook Air 2015.