fix: ensure non-existent file doesn't exist
The test can_handle_file_not_found bets on the non-existence of test.s,
a very common file name.

This commit changes it so it uses a file
that's much more likely to not exist and also asserts that it doesn't
exist.

Not the perfect fix, but will do.

Fixes #500.