fix(dev): allow actions requiring secrets from safe external PRs (#1282)
* fix(dev): allow actions requiring secrets from safe external PRs

By default, secrets are not available from workflows triggered
by external PRs, which go through the pull_request event.

This PR adds a way to execute workflows for those PRs via
the (more dangerous) event pull_request_target, which runs
in the context of the base branch and provides secrets, and
in this case is triggered by the "safe to test" label, which can
only be set by maintainers.

* fix: only run the fallback step if 'deployment' did not succeed